fix: fixed underlying issue which was that party selection did not work properly
